Update __init__.py
two changes in order to drive sprites with bones. The main idea is to use puppet rig as linked group in shots, and still have control on sprites slot index. Another change to update in realtime a slot index when a driving bone is used.
Use at your own risk